REFACTOR: move email sending functions to email.py
